refactor events and use of singletone class
