+++ dagrunner/dagrunner.cfg
@@ -1,7 +1,7 @@
 [dagrunner_visualisation]
 #title=dummy
 #enabled=True
-#collapse_properties=leadtime
+#collapse_properties=leadtime,cycle_offset
 #backend=mermaid
 #group_by=diagnostic
 #label_by=step
